src/ch7/diagram-agent/requirements.txt
Flask>=3.0.0
pywebview>=5.1
litellm>=1.50.0
